requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.7
aiohttp-retry==2.9.1
aiosignal==1.3.1
alembic==1.13.3
annotated-types==0.7.0
anyio==4.6.2.post1
APScheduler==3.9.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
async-lru==2.0.5
attrs==24.2.0
babel==2.17.0
bcrypt==4.2.0
beautifulsoup4==4.13.3
bleach==6.2.0
blinker==1.7.0
cached-property==2.0.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cryptography==43.0.1
cycler==0.12.1
debugpy==1.8.13
decorator==5.2.1
defusedxml==0.7.1
distlib==0.3.8
Django==5.1.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
executing==2.0.1
fastapi==0.115.5
fastapi-cli==0.0.7
fastjsonschema==2.21.1
filelock==3.15.4
Flask==2.3.2
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.6.0
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-MySQLdb==2.0.0
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.2.1
fonttools==4.56.0
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2025.2.0
funcs==0.11.0
greenlet==3.0.3
gunicorn==20.1.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
icecream==2.1.3
idna==3.10
imageio==2.37.0
importlib-metadata==5.0.0
inflection==0.5.1
iniconfig==2.0.0
ipykernel==6.29.5
ipython==9.0.2
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.2
Jinja2==3.1.3
joblib==1.4.2
jose==1.0.0
json5==0.10.0
jsonify==0.5
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
jwt==1.3.1
kiwisolver==1.4.8
lazy_loader==0.4
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
mysql-connector==2.2.9
mysql-connector-python==9.2.0
mysqlclient==2.2.7
named==1.4.2
nba_api==1.8.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
notebook==7.3.3
notebook_shim==0.2.4
numpy==1.26.4
opencv-python==4.11.0.86
optional-django==0.3.0
orjson==3.10.15
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
passlib==1.7.4
patool==3.1.0
pillow==10.4.0
pipenv==2024.1.0
platformdirs==4.2.2
pluggy==1.5.0
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.2.0
psutil==7.0.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.10.6
pydantic-extra-types==2.10.2
pydantic-settings==2.7.1
pydantic_core==2.27.2
Pygments==2.18.0
PyJWT==1.7.1
PyMySQL==1.1.1
pyodbc==5.2.0
pyparsing==3.2.1
PySocks==1.7.1
pytest==8.3.3
pytest-flask==1.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-json-logger==3.3.0
python-multipart==0.0.20
pytz==2022.5
pytz-deprecation-shim==0.1.0.post0
pywin32==310
pywinpty==2.0.15
PyYAML==6.0.2
pyzmq==26.3.0
react==4.3.0
referencing==0.36.2
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rich-toolkit==0.13.2
rpds-py==0.23.1
rsa==4.9
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
selenium==4.29.0
Send2Trash==1.8.3
setuptools==75.1.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
solus==1.2.2
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.25
sqlalchemy-orm==1.2.10
sqlparse==0.5.1
stack-data==0.6.3
starlette==0.41.3
sympy==1.13.1
terminado==0.18.1
threadpoolctl==3.6.0
tifffile==2025.3.13
tinycss2==1.4.0
torch==2.6.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
trio==0.29.0
trio-websocket==0.12.2
twilio==9.4.4
typer==0.15.1
types-python-dateutil==2.9.0.20241206
typing-aliases==1.10.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==4.2
ujson==5.10.0
uri-template==1.3.0
urllib3==2.2.3
uvicorn==0.32.1
virtualenv==20.26.3
watchfiles==1.0.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.1
Werkzeug==2.3.7
widgetsnbextension==4.0.13
wraps==0.15.0
wsproto==1.2.0
WTForms==3.1.2
yarl==1.18.0
zipp==3.10.0
